% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgp_birm_rs.R
\name{dgp_birm_rs}
\alias{dgp_birm_rs}
\title{Generate a dataset}
\usage{
dgp_birm_rs(
  n = 2000,
  item_n = "auto",
  theta_n = 1,
  var_thetas = "auto",
  var_ers = 1,
  var_ars = 1,
  cor_thetas = "auto",
  x_num = "auto",
  cor_ers = "auto",
  include_ERS = TRUE,
  include_ARS = TRUE,
  seed = sample(1:1e+09, 1)
)
}
\arguments{
\item{n}{Number of observations.}

\item{item_n}{Number of items. For multiple traits, this is a vector with the number of items per trait. When "auto", will be set to 10 for all traits.}

\item{theta_n}{Number of latent traits (1 to 3).}

\item{var_thetas}{variances of latent traits. Defaults to 1 for all traits.}

\item{var_ers}{variance of ERS parameters. Defaults to 1.}

\item{var_ars}{variance of ARS parameters. Defaults to 1.}

\item{cor_thetas}{correlation between latent traits. When modeling three traits, the order of correlations is: cor(theta1, theta2), cor(theta1, theta3), cor(theta2, theta3). Defaults to runif((theta_n*(theta_n - 1))/2, -0.4, 0.4).}

\item{x_num}{number of reverse-coded items per theta. Defaults to floor(item_n/2).}

\item{cor_ers}{correlation between ERS and thetas. When modeling three traits, the order of correlations is: cor(theta1, ERS), cor(theta2, ERS), cor(theta3, ERS). Defaults to runif(theta_n, -0.3, 0.3).}
}
\value{
A list with two dataframes. One with the item parameters and one with the person parameters and simulated data.
}
\description{
Generates item and person parameters and then generates a datset with these parameters.
}
\details{
Generate item and person parameters and then generates a datset with these parameters.
}
\examples{
dfs_list <- dgp_birm_rs(n = 2000, item_n = 10, theta_n = 1, var_thetas = rep(1, theta_n), var_ers = 1, var_ars = 1, x_num = floor(item_n/2), cor_ers = 0.2)
}
