% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_stan.R
\name{fit_stan}
\alias{fit_stan}
\title{Fit a BIRM-RS model}
\usage{
fit_stan(
  data,
  stan_model,
  theta_n,
  x_vec,
  T_vec = "auto",
  iter = 8000,
  warmup = 4000,
  chains = 6,
  seed = sample(1:1e+09, 1),
  adapt_delta = 0.9,
  prefix = "observed",
  ars_prior = 0.9,
  init_vals = TRUE
)
}
\arguments{
\item{data}{Data frame containing the observed data.}

\item{stan_model}{Path to the Stan model file.}

\item{theta_n}{Number of latent traits.}

\item{x_vec}{Vector indicating reverse-scored items (1 = normally scored, -1 = reverse-scored).}

\item{T_vec}{Vector specifying the number of items per trait. Defaults to "auto", which sets it to the total number of items divided by theta_n.}

\item{iter}{Number of iterations (excluding warmup).}

\item{warmup}{Number of warmup iterations.}

\item{chains}{Number of MCMC chains.}

\item{seed}{Random seed for reproducibility.}

\item{adapt_delta}{Target acceptance rate for step size adaptation in Stan. Defaults to 0.9.}

\item{prefix}{Optional prefix indicating which variables to analyze. Default is "observed". If NULL, all columns are used.}

\item{ars_prior}{Prior for the ARS parameter. Default is 0.9.}

\item{init_vals}{Logical; indicates whether to use initial values. Default is FALSE.}
}
\value{
A stanfit object representing the fitted model.
}
\description{
Fits a BIRM-RS model to the provided dataset.
}
\examples{
results <- fit_stan(df, stan_model = "stan/ARS_ERS.stan", theta_n = 2, 
                    x_vec = c(rep(1, 4), rep(-1, 4)), iter = 8000, warmup = 4000, chains = 6, 
                    seed = sample(1:1e9, 1), adapt_delta = 0.9, prefix = "observed", init_vals = TRUE)
}
