% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_stan.R
\name{fit_stan}
\alias{fit_stan}
\title{Fit a RS-BIRM model}
\usage{
fit_stan(
  data,
  stan_model,
  theta_n,
  x_vec,
  T_vec = "auto",
  iter = 8000,
  warmup = 4000,
  chains = 6,
  seed = sample(1:1e+09, 1),
  adapt_delta = 0.9,
  prefix = "observed",
  ars_prior = 0.9,
  init_vals = TRUE
)
}
\arguments{
\item{data}{data frame with the data}

\item{stan_model}{path to the stan model file}

\item{theta_n}{number of latent traits}

\item{x_vec}{vector of item codings}

\item{T_vec}{vector of item numbers per trait}

\item{iter}{number of iterations}

\item{warmup}{number of warmup iterations}

\item{chains}{number of chains}

\item{seed}{random seed}

\item{adapt_delta}{adaptation delta}

\item{prefix}{Should only variables with the prefix be analyzed? Default is NULL}
}
\value{
a stanfit object
}
\description{
Fits a RS-BIRM model to the data.
}
\examples{
results <- fit_stan(df, stan_model = "stan/ARS_ERS.stan", theta_n = 2, 
                    x_vec = c(rep(1, 4), rep(-1, 4)), iter = 8000, warmup = 4000, chains = 6, 
                    seed = sample(1:1e9, 1), adapt_delta = 0.9, prefix = "observed", init_vals = TRUE)
}
