% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/one_simulation.R
\name{one_simulation}
\alias{one_simulation}
\title{Simulates data, fits a model and summarizes the fit}
\usage{
one_simulation(shorten = TRUE, ...)
}
\arguments{
\item{...}{arguments passed to \code{\link{dgp_birm_rs}}, \code{\link{fit_stan}} and \code{\link{summarize_fit}}}
}
\description{
Summarizes a fit object by computing means, standard deviations, minima and maxima of the posterior distributions of the model parameters, and the root mean squared error of the model parameter estimates.
}
