% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_fun.R
\name{sim_fun}
\alias{sim_fun}
\title{Run Simulations in Parallel with Progress Reporting}
\usage{
sim_fun(
  sim_grid,
  results_path,
  workers = "half",
  save_all = TRUE,
  parallel_type = "multisession"
)
}
\arguments{
\item{sim_grid}{A data frame containing the grid of simulation parameters.}

\item{results_path}{A file path to save individual simulation results.}

\item{workers}{Number of parallel workers to use. Can be numeric, or "half"/"quarter".}

\item{save_all}{Logical; whether to save all individual simulation result objects to disk.}
}
\value{
A data frame containing a summary of all simulations.
Failed simulations are padded to match the structure of successful ones.
}
\description{
This function runs simulations in parallel based on the given simulation grid.
It compiles Stan models, executes simulations using parallel workers,
handles warnings and errors, and saves results.
}
