% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_fit.R
\name{summarize_fit}
\alias{summarize_fit}
\title{Summarize a fit object from \code{\link{fit_stan}}}
\usage{
summarize_fit(fit, df_list, theta_n = "auto")
}
\arguments{
\item{fit}{fit-object from \code{\link{fit_stan}}}

\item{df_list}{List containing the data used to fit the model and the items used to fit the model}

\item{theta_n}{Number of latent traits}
}
\value{
A list with the summaries and the root mean squared errors
}
\description{
Summarizes a fit object by computing means, standard deviations, minima and maxima of the posterior distributions of the model parameters, and the root mean squared error of the model parameter estimates.
}
