% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_fit.R
\name{summarize_fit}
\alias{summarize_fit}
\title{Summarize a fit object from \code{\link{fit_stan}}}
\usage{
summarize_fit(fit, df_list, theta_n = "auto")
}
\arguments{
\item{fit}{Fit object from \code{\link{fit_stan}}.}

\item{df_list}{List containing the data used to fit the model and the items used to fit the model.}

\item{theta_n}{Number of latent traits. If not specified, it will be automatically inferred from the column names of \code{df_list$df}.}
}
\value{
A list containing summary statistics and the RMSEs of the model parameters.
}
\description{
Summarizes the results from a fit object, including means, standard deviations, minima, and maxima of the posterior distributions of the model parameters, as well as the root mean squared error (RMSE) of the model parameter estimates.
}
